♻️  macros: Refactor `proxy` macro
